Maze Generator Class
This file is where the maze is created. The maze is created by initially making an n x n grid of cells. Each of these Cells begins as its own set. A random integer is selected and the corresponding wall is removed. When a wall is removed the 2 cells which shared the removed wall are combined into the same set.